requirements.txt
flask
python-dotenv
sentence-transformers
faiss-cpu
pydantic
pypdf
beautifulsoup4
requests
google-search-results>=2.4.2
# LangChain
langchain>=0.2.14
langchain-openai>=0.1.22
langchain-community>=0.2.12